+++ Datadog/IntegrationUnitTests/Trace/HeadBasedSamplingTests.swift
@@ -110,7 +110,7 @@ class HeadBasedSamplingTests: XCTestCase {
             firstPartyHostsTracing: .trace(hosts: ["foo.com"], sampleRate: distributedTraceSampling)
         )
         Trace.enable(with: traceConfig, in: core)
-        URLSessionInstrumentation.enable(with: .init(delegateClass: InstrumentedSessionDelegate.self), in: core)
+        URLSessionInstrumentation.trackMetrics(with: .init(delegateClass: InstrumentedSessionDelegate.self), in: core)
 
         // When
         let request = try sendURLSessionRequest(to: "https://foo.com/request", using: InstrumentedSessionDelegate())
@@ -149,7 +149,7 @@ class HeadBasedSamplingTests: XCTestCase {
             firstPartyHostsTracing: .trace(hosts: ["foo.com"], sampleRate: distributedTraceSampling, traceControlInjection: .all)
         )
         Trace.enable(with: traceConfig, in: core)
-        URLSessionInstrumentation.enable(with: .init(delegateClass: InstrumentedSessionDelegate.self), in: core)
+        URLSessionInstrumentation.trackMetrics(with: .init(delegateClass: InstrumentedSessionDelegate.self), in: core)
 
         // When
         let request = try sendURLSessionRequest(to: "https://foo.com/request", using: InstrumentedSessionDelegate())
@@ -181,7 +181,7 @@ class HeadBasedSamplingTests: XCTestCase {
             firstPartyHostsTracing: .trace(hosts: ["foo.com"], sampleRate: distributedTraceSampling)
         )
         Trace.enable(with: traceConfig, in: core)
-        URLSessionInstrumentation.enable(with: .init(delegateClass: InstrumentedSessionDelegate.self), in: core)
+        URLSessionInstrumentation.trackMetrics(with: .init(delegateClass: InstrumentedSessionDelegate.self), in: core)
 
         // When
         let span = Tracer.shared(in: core).startSpan(operationName: "active.span").setActive()
@@ -229,7 +229,7 @@ class HeadBasedSamplingTests: XCTestCase {
             firstPartyHostsTracing: .trace(hosts: ["foo.com"], sampleRate: distributedTraceSampling, traceControlInjection: .all)
         )
         Trace.enable(with: traceConfig, in: core)
-        URLSessionInstrumentation.enable(with: .init(delegateClass: InstrumentedSessionDelegate.self), in: core)
+        URLSessionInstrumentation.trackMetrics(with: .init(delegateClass: InstrumentedSessionDelegate.self), in: core)
 
         // When
         let span = Tracer.shared(in: core).startSpan(operationName: "active.span").setActive()